Update and rename Mandelbrot.v1.rkt to Mandelbrot.v3.rkt
Major changes: 
mandelbrot? -> mandelbrot-number
Changed from a predicate to a function that returns a number representing a Mandelbrot escape number. This change will make allow for rendered images to have multiple colors representing the escape number of a given pixel.

Mandelbrot-Image
Because the Mandelbrot set is symmetric I only need to render the top half and I can get the bottom half for free by flipping the rendered image.

Minor changes:
draw function has been changed to allow for different colored dots to be placed on the image. Orientation of the dots have also been flipped so that dot position is relative to top left corner.

Removed guide marks from background image when it was decided to only render half of the image.