Update and rename Mandelbrot.v4.rkt to Mandelbrot.v5.rkt
Major changes:
Changed functions to use float values instead of complex numbers. This decreased the time required to generate the mandelbrot escape numbers by an order of magnitude and allowed for the quick generation of scenes over 1000 pixels in size.

Before this change it took 20 seconds to generate the list of escape numbers required for a 300 by 200 image with an iteration depth of 9. This calculation can now by finished in 0.24 seconds.

Images of size 1800 by 1200 with iteration depth of 120 can now be generated in 40 seconds.

Minor change:
Sense I can now calculate iteration depths greater than 10 quickly, I changed the way that dot colors are calculated to better reflect their depth